+++ NEWS.md
@@ -1,5 +1,7 @@
 # euroaq (development version)
 
+* `get_eea_parquet_files()` and family now use `dynamic = TRUE` by default. When `dynamic = FALSE` the API calls will route to the 'classic' endpoints.
+
 * `get_eea_ddb_extension_info()` and `get_eea_ddb_vars()` have been removed as these are no longer endpoints available from the EEA API.
 
 * `import_eea_pollutants()` has been fixed. This was broken due to changes in the API return values.